+++ example/app/main.aot.ts
@@ -1,4 +1,7 @@
+import { enableProdMode } from "@angular/core";
 import { platformBrowser }    from "@angular/platform-browser";
 import { AppModuleNgFactory } from "../aot/example/app/app.module.ngfactory";
+
+enableProdMode();
 
 platformBrowser().bootstrapModuleFactory(AppModuleNgFactory);