+++ modules/core/src/service/dynamic-form.service.ts
@@ -137,9 +137,9 @@ export class DynamicFormService {
     }
 
 
-    createFormGroup(group: DynamicFormControlModel[], groupExtra: {[key: string]: any} | null = null): FormGroup {
+    createFormGroup(group: DynamicFormControlModel[], groupExtra: { [key: string]: any } | null = null): FormGroup {
 
-        let formGroup: {[id: string]: AbstractControl;} = {};
+        let formGroup: { [id: string]: AbstractControl; } = {};
 
         group.forEach(model => {
 
@@ -182,17 +182,50 @@ export class DynamicFormService {
                         groupModel: DynamicFormControlModel[] | DynamicFormGroupModel,
                         ...controlModels: DynamicFormControlModel[]): void {
 
+        if (groupModel instanceof DynamicFormGroupModel) {
+
+            this.insertFormGroupControl(groupModel.size(), formGroup, groupModel, ...controlModels);
+
+        } else {
+
+            let formModel = groupModel as DynamicFormControlModel[];
+            this.insertFormGroupControl(formModel.length, formGroup, formModel, ...controlModels);
+        }
+    }
+
+
+    moveFormGroupControl(index: number,
+                         step: number,
+                         groupModel: DynamicFormControlModel[] | DynamicFormGroupModel): void {
+
+        if (groupModel instanceof DynamicFormGroupModel) {
+
+            groupModel.move(index, step);
+
+        } else {
+
+            let formModel = groupModel as DynamicFormControlModel[];
+            formModel.splice(index + step, 0, ...formModel.splice(index, 1));
+        }
+    }
+
+
+    insertFormGroupControl(index: number,
+                           formGroup: FormGroup,
+                           groupModel: DynamicFormControlModel[] | DynamicFormGroupModel,
+                           ...controlModels: DynamicFormControlModel[]): void {
+
         let controls = this.createFormGroup(controlModels).controls;
 
-        Object.keys(controls).forEach((controlName, index) => {
+        Object.keys(controls).forEach((controlName, idx) => {
 
-            let controlModel = controlModels[index];
+            let controlModel = controlModels[idx];
 
             if (groupModel instanceof DynamicFormGroupModel) {
-                groupModel.add(controlModel);
+                groupModel.insert(index, controlModel);
 
             } else {
-                (groupModel as DynamicFormControlModel[]).push(controlModel);
+                (groupModel as DynamicFormControlModel[]).splice(index, 0, controlModel);
             }
 
             formGroup.addControl(controlName, controls[controlName]);
@@ -202,7 +235,7 @@ export class DynamicFormService {
 
     removeFormGroupControl(index: number,
                            formGroup: FormGroup,
-                           groupModel: DynamicFormControlModel[] | DynamicFormGroupModel) {
+                           groupModel: DynamicFormControlModel[] | DynamicFormGroupModel): void {
 
         if (groupModel instanceof DynamicFormGroupModel) {
 
@@ -229,27 +262,45 @@ export class DynamicFormService {
     }
 
 
-    removeFormArrayGroup(index: number, formArray: FormArray, model: DynamicFormArrayModel): void {
-
-        formArray.removeAt(index);
-        model.removeGroup(index);
-    }
-
-
     moveFormArrayGroup(index: number, step: number, formArray: FormArray, model: DynamicFormArrayModel): void {
 
         let newIndex = index + step,
-            formControl = formArray.at(index);
+            moveUp = step >= 0;
 
         if ((index >= 0 && index < model.size) && (newIndex >= 0 && newIndex < model.size)) {
 
-            formArray.setControl(index, formArray.at(newIndex));
-            formArray.setControl(newIndex, formControl);
+            let moving = [];
+
+            for (let i = moveUp ? index : newIndex; i <= (moveUp ? newIndex : index); i++) {
+                moving.push(formArray.at(i));
+            }
+
+            moving.forEach((formControl, idx) => {
+
+                let position;
+
+                if (moveUp) {
+                    position = idx === 0 ? newIndex : index + idx - 1;
+
+                } else {
+                    position = idx === moving.length - 1 ? newIndex : newIndex + idx + 1;
+                }
+
+                formArray.setControl(position, formControl);
+            });
+
             model.moveGroup(index, step);
 
         } else {
             throw new Error(`form array group cannot be moved due to index or new index being out of bounds`);
         }
+    }
+
+
+    removeFormArrayGroup(index: number, formArray: FormArray, model: DynamicFormArrayModel): void {
+
+        formArray.removeAt(index);
+        model.removeGroup(index);
     }