+++ modules/ui-ionic/src/dynamic-form-ionic.component.ts
@@ -55,7 +55,7 @@ export class DynamicFormIonicComponent extends DynamicFormControlComponent {
     @Input()set controlGroup(group: FormGroup) {
         this.group = group;
         console.warn("[controlGroup] is deprecated. Use [group] instead.");
-    };
+    }
 
     @Input() group: FormGroup;
     @Input() hasErrorMessaging: boolean = false;